check if model id is None or empty

+++ sdk/formrecognizer/azure-ai-formrecognizer/azure/ai/formrecognizer/_form_training_client.py
@@ -163,6 +163,9 @@ class FormTrainingClient(object):
                 :caption: Delete a custom model.
         """
 
+        if not model_id:
+            raise ValueError("model_id cannot be None or empty.")
+
         self._client.delete_custom_model(
             model_id=model_id,
             error_map=error_map,
@@ -236,6 +239,10 @@ class FormTrainingClient(object):
                 :dedent: 8
                 :caption: Get a custom model with a model ID.
         """
+
+        if not model_id:
+            raise ValueError("model_id cannot be None or empty.")
+
         response = self._client.get_custom_model(model_id=model_id, include_keys=True, error_map=error_map, **kwargs)
         return CustomFormModel._from_generated(response)
